+++ GRDB/QueryInterface/SQL/Table.swift
@@ -1186,6 +1186,88 @@ extension Table {
     {
         all().joining(required: association)
     }
+    
+    /// Creates a request which appends *columns of an associated record* to
+    /// the columns of the table.
+    ///
+    ///     let playerTable = Table("player")
+    ///     let teamTable = Table("team")
+    ///     let playerTeam = playerTable.belongsTo(teamTable)
+    ///
+    ///     // SELECT player.*, team.color
+    ///     // FROM player LEFT JOIN team ...
+    ///     let teamColor = playerTeam.select(Column("color")
+    ///     let request = playerTable.annotated(withOptional: teamColor))
+    ///
+    /// This method performs the same SQL request as `including(optional:)`.
+    /// The difference is in the shape of Decodable records that decode such
+    /// a request: the associated columns can be decoded at the same level as
+    /// the main record:
+    ///
+    ///     struct PlayerWithTeamColor: FetchableRecord, Decodable {
+    ///         var player: Player
+    ///         var color: String?
+    ///     }
+    ///     let players = try dbQueue.read { db in
+    ///         try request
+    ///             .asRequest(of: PlayerWithTeamColor.self)
+    ///             .fetchAll(db)
+    ///     }
+    ///
+    /// Note: this is a convenience method. You can build the same request with
+    /// `TableAlias`, `annotated(with:)`, and `joining(optional:)`:
+    ///
+    ///     let teamAlias = TableAlias()
+    ///     let request = playerTable
+    ///         .annotated(with: teamAlias[Column("color")])
+    ///         .joining(optional: playerTeam.aliased(teamAlias))
+    public func annotated<A: Association>(withOptional association: A)
+    -> QueryInterfaceRequest<RowDecoder>
+    where A.OriginRowDecoder == RowDecoder
+    {
+        all().annotated(withOptional: association)
+    }
+    
+    /// Creates a request which appends *columns of an associated record* to
+    /// the columns of the table.
+    ///
+    ///     let playerTable = Table("player")
+    ///     let teamTable = Table("team")
+    ///     let playerTeam = playerTable.belongsTo(teamTable)
+    ///
+    ///     // SELECT player.*, team.color
+    ///     // FROM player JOIN team ...
+    ///     let teamColor = playerTeam.select(Column("color")
+    ///     let request = playerTable.annotated(withRequired: teamColor))
+    ///
+    /// This method performs the same SQL request as `including(required:)`.
+    /// The difference is in the shape of Decodable records that decode such
+    /// a request: the associated columns can be decoded at the same level as
+    /// the main record:
+    ///
+    ///     struct PlayerWithTeamColor: FetchableRecord, Decodable {
+    ///         var player: Player
+    ///         var color: String
+    ///     }
+    ///     let players = try dbQueue.read { db in
+    ///         try request
+    ///             .asRequest(of: PlayerWithTeamColor.self)
+    ///             .fetchAll(db)
+    ///     }
+    ///
+    /// Note: this is a convenience method. You can build the same request with
+    /// `TableAlias`, `annotated(with:)`, and `joining(required:)`:
+    ///
+    ///     let teamAlias = TableAlias()
+    ///     let request = playerTable
+    ///         .annotated(with: teamAlias[Column("color")])
+    ///         .joining(required: playerTeam.aliased(teamAlias))
+    public func annotated<A: Association>(withRequired association: A)
+    -> QueryInterfaceRequest<RowDecoder>
+    where A.OriginRowDecoder == RowDecoder
+    {
+        all().annotated(withRequired: association)
+    }
 }
 
 // MARK: - Association Aggregates